+++ packages/web/src/components/gcds-pagination/gcds-pagination.tsx
@@ -2,6 +2,7 @@ import { Component, Element, Prop, Event, EventEmitter, State, Watch, Host, h } 
 
 import { assignLanguage } from '../../utils/utils';
 import I18N from './i18n/i18n';
+import { constructHref, constructClasses } from './utils/render';
 
 interface PaginationUrl {
   queryStrings: Object;
@@ -100,236 +101,13 @@ export class GcdsPagination {
   }
 
   /**
-   * Function to constuct href attribute from url object
-   * Also looks for ::offset and ::match to increment query string values
-   */
-  private constructHref(page: number, end?: "next" | "previous" | null) {
-    var fragment = "";
-    var qs = "";
-
-    if (this.url) {
-      var count = 0;
-      for (const key in this.url.queryStrings) {
-        var queryKey = key;
-        var queryValue = this.url.queryStrings[key];
-
-        if (key.includes("::")) {
-          var incrementer = key.split("::")[1];
-          var regExp = /\{\{([^)]+)\}\}/;
-          var matches = regExp.exec(this.url.queryStrings[key]);
-
-          // Offeset numbers
-          if (incrementer == "offset") {
-            var pageNumber = page;
-
-            if (end == "next") {
-              ++pageNumber;
-            }
-            else if (end == "previous") {
-              --pageNumber;
-            }
-
-            queryValue = matches ?
-              this.url.queryStrings[key].replace(`{{${matches[1]}}}`, `${((pageNumber-1) * Number(matches[1]))}`)
-            :
-              ((pageNumber-1) * this.url.queryStrings[key]);
-
-            queryKey = queryKey.replace("::offset", "");
-          }
-
-          // Match page number
-          if (incrementer == "match") {
-            var pageNumber = page;
-
-            if (end == "next") {
-              ++pageNumber;
-            }
-            else if (end == "previous") {
-              --pageNumber;
-            }
-
-            queryValue = matches ?
-              this.url.queryStrings[key].replace(`{{${matches[1]}}}`, `${(pageNumber * Number(matches[1]))}`)
-            :
-              (pageNumber * this.url.queryStrings[key]);
-
-            queryKey = queryKey.replace("::match", "");
-          }
-        }
-
-        if (count == 0) {
-          qs += `?${queryKey}=${queryValue}`
-        } else {
-          qs += `&${queryKey}=${queryValue}`
-        }
-        ++count;
-      }
-
-      if (this.url.fragment) {
-        fragment = `#${this.url.fragment}`;
-      }
-
-    }
-
-    var href = `${qs}${fragment}`;
-
-    return href;
-  }
-
-  /**
-   * Function to constuct classes to help with mobile sizing
-   */
-  private constructClasses(page: number, current: number, total: number) {
-    if (total <= 6) {
-
-    }
-    else if (total < 10 && (current == 1 || current == total)) {
-      switch(current - page) {
-        case 4:
-        case -4:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-360";
-          }
-          break;
-        case 5:
-        case -5:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-420";
-          }
-          break;
-        case 6:
-        case -6:
-        case 7:
-        case -7:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-488";
-          }
-          break;
-      }
-    }
-    else if (total < 10 && (current == 2 || current == (total-1))) {
-      switch(current - page) {
-        case 3:
-        case -3:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-360";
-          }
-          break;
-        case 4:
-        case -4:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-420";
-          }
-          break;
-        case 5:
-        case -5:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-488";
-          }
-          break;
-      }
-    }
-    else if (total < 10 && (current > 2 && current < (total-1))) {
-      switch(current - page) {
-        case 2:
-        case -2:
-            if (page != 1 && page != total ){
-              return "gcds-pagination-list-breakpoint-420";
-            }
-            break;
-        case 3:
-        case -3:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-460";
-          }
-          break;
-        case 4:
-        case -4:
-        case 5:
-        case -5:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-488";
-          }
-          break;
-      }
-    }
-    else if (current == 2 || current == (total-1)) {
-      switch(current - page) {
-        case 3:
-        case -3:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-360";
-          }
-          break;
-        case 4:
-        case -4:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-420";
-          }
-          break;
-        case 5:
-        case -5:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-488";
-          }
-          break;
-      }
-    }
-    else if (current < 3 || current > (total-2)) {
-      switch(current - page) {
-        case 4:
-        case -4:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-420";
-          }
-          break;
-        case 5:
-        case -5:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-460";
-          }
-          break;
-        case 6:
-        case -6:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-488";
-          }
-          break;
-      }
-    }
-    else {
-      switch(current - page) {
-        case 2:
-        case -2:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-420";
-          }
-          break;
-        case 3:
-        case -3:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-460";
-          }
-          break;
-        case 4:
-        case -4:
-          if (page != 1 && page != total ){
-            return "gcds-pagination-list-breakpoint-488";
-          }
-          break;
-      }
-    }
-    return "";
-  }
-
-  /**
    * Function to constuct <li> and <a> tags for display="list" pagination
    */
   private configurePaginationStep(page: number, end?: "next" | "previous" | null, mobile?: boolean) {
 
     var linkAttrs = {
       onClick: (e) => this.onPageChange(e),
-      href: this.url ? this.constructHref(page, end) : "javascript:void(0)",
+      href: this.url ? constructHref(this.el, page, end) : "javascript:void(0)",
       ariaLabel: !end ?
         I18N[this.lang].pageNumberOf.replace('{#}', page).replace('{label}', this.label)
       :
@@ -350,7 +128,7 @@ export class GcdsPagination {
       return(<li><a  {...linkAttrs} class={!mobile ? "gcds-pagination-end-button" : "gcds-pagination-end-button-mobile"}><gcds-icon class="gcds-pagination-arrow-left" name="arrow-left"></gcds-icon>{I18N[this.lang].previous}</a></li>);
     }
     else {
-      return (<li class={this.constructClasses(page, this.currentPage, this.totalPages)}><a {...linkAttrs}>{page}</a></li>);
+      return (<li class={constructClasses(page, this.currentPage, this.totalPages)}><a {...linkAttrs}>{page}</a></li>);
     }
 
   }
@@ -423,12 +201,7 @@ export class GcdsPagination {
       else if (i == 1 || i == this.totalPages) {
         this.listitems.push(this.configurePaginationStep(i));
       }
-
-      // // Second page when within 4
-      // else if (this.currentPage <= 4 && i == 2) {
-      //   this.listitems.push(this.configurePaginationStep(i));
-      // }
-      // Up to 5 when less than 4
+      // Up to 7 when less than 6
       else if (this.currentPage < 6 && i <= 7) {
         this.listitems.push(this.configurePaginationStep(i));
         if (i==2 && (this.currentPage == 4 || this.currentPage == 5)) {
@@ -447,10 +220,6 @@ export class GcdsPagination {
           this.listitems.push(<li class={`gcds-pagination-list-mobile-ellipses-${breakpoint}`}><span class="gcds-pagination-list-ellipses">...</span></li>);
         }
       }
-      // // Second last page when within 4
-      // else if (this.currentPage > this.totalPages - 4 && i == this.totalPages - 1) {
-      //   this.listitems.push(this.configurePaginationStep(i));
-      // }
       // Show last numbers within 5
       else if (this.currentPage > this.totalPages - 5 && i >= this.totalPages - 6) {
         if (i == this.totalPages - 1 && (this.currentPage == this.totalPages - 3 || this.currentPage == this.totalPages - 4)) {
@@ -509,7 +278,7 @@ export class GcdsPagination {
   }
 
   render() {
-    const { display, label, previousHref, previousLabel, nextHref, nextLabel } = this;
+    const { display, label, previousHref, previousLabel, nextHref, nextLabel, lang } = this;
 
     return (
       <Host
@@ -536,12 +305,12 @@ export class GcdsPagination {
             <li>
               <a
                 href={previousHref}
-                aria-label={`Previous page: ${previousLabel}`}
+                aria-label={`${I18N[lang].nextPage}: ${previousLabel}`}
                 onClick={(e) => this.onPageChange(e)}
               >
                 <gcds-icon class="gcds-pagination-arrow-left" name="arrow-left"></gcds-icon>
                 <b>
-                  Previous
+                  {I18N[lang].previous}
                 </b>
                 <span>
                   {previousLabel}
@@ -551,12 +320,12 @@ export class GcdsPagination {
             <li>
               <a
                 href={nextHref}
-                aria-label={`Next page: ${nextLabel}`}
+                aria-label={`${I18N[lang].nextPage}: ${nextLabel}`}
                 onClick={(e) => this.onPageChange(e)}
               >
                 <gcds-icon class="gcds-pagination-arrow-right" name="arrow-right"></gcds-icon>
                 <b>
-                  Next
+                  {I18N[lang].next} 
                 </b>
                 <span>
                   {nextLabel}
