+++ packages/web/src/index.html
@@ -52,13 +52,12 @@
         total-pages="20"
         current-page="10"
         id="page"
-        lang="fr"
       >
       </gcds-pagination>
 
       <br/>
 
-      <gcds-pagination
+     <gcds-pagination
         display="list"
         label="Search results"
